+++ R/adsl.R
@@ -126,5 +126,5 @@
 #' @keywords datasets adsl
 #' @name adsl
 #' @examples
-#' head(data("adsl"))
+#'  head(data("adsl"))
 "adsl"