+++ NEWS.md
@@ -18,6 +18,7 @@
 - Releveled TRT01P and TRT01A (now Xanomeline Low Dose -> Xanomeline High Dose -> Placebo)
 - Modified "General disorders and administration site conditions" to "Gastrointestinal disorders" in ADEX
 - Modified ADVS on ORTHYP, ORTHYPS and  ORTHYPD parameters for TSFVIT04
+- Modified RESCRNFL and assigned 01-701-1240 to Subject refused to sign informed consent
 
 ## [0.0.3] - 2026-01-27